Restart refactor
